feature: experimentally support configured I2C freq
This feature is useful when you need maximum possible frequency with
multiple I2C slaves.

Not all I2C libraries offer functions to implement this feature. Brzo
I2C is the only exception, which requires SCL freq to be set when
starting an I2C transaction.

With esp-idf, it is possible but requires the library to be tightly
coupled with its I2C component (it does not have an easy way to set or
get SCL freq like i2c_get_clk_speed()).

arduino-esp8266 does not implement Wire.getClock() (its Wire.h is very
old).

As such, disabled by default.